+++ src/e2e/codegen/mod.rs
@@ -34,6 +34,7 @@ pub mod kotlin;
 pub mod kotlin_android;
 pub mod php;
 pub mod php_ext;
+mod presentation;
 pub mod python;
 pub mod r;
 pub mod recipe;